Initial draft of splitting quant util.
