+++ src/Makefile.am
@@ -406,8 +406,12 @@ libcvc4_la_SOURCES = \
 	theory/quantifiers/model_engine.h \
 	theory/quantifiers/quant_conflict_find.cpp \
 	theory/quantifiers/quant_conflict_find.h \
+	theory/quantifiers/quant_epr.cpp \
+	theory/quantifiers/quant_epr.h \
 	theory/quantifiers/quant_equality_engine.cpp \
 	theory/quantifiers/quant_equality_engine.h \
+	theory/quantifiers/quant_relevance.cpp \
+	theory/quantifiers/quant_relevance.h \
 	theory/quantifiers/quant_split.cpp \
 	theory/quantifiers/quant_split.h \
 	theory/quantifiers/quant_util.cpp \
@@ -422,6 +426,8 @@ libcvc4_la_SOURCES = \
 	theory/quantifiers/rewrite_engine.h \
 	theory/quantifiers/sygus_grammar_cons.cpp \
 	theory/quantifiers/sygus_grammar_cons.h \
+	theory/quantifiers/sygus_invariance.cpp \
+	theory/quantifiers/sygus_invariance.h \
 	theory/quantifiers/sygus_process_conj.cpp \
 	theory/quantifiers/sygus_process_conj.h \
 	theory/quantifiers/symmetry_breaking.cpp \
