+++ src/theory/sep/theory_sep.cpp
@@ -25,6 +25,7 @@
 #include "options/smt_options.h"
 #include "smt/logic_exception.h"
 #include "theory/quantifiers_engine.h"
+#include "theory/quantifiers/quant_epr.h"
 #include "theory/quantifiers/term_database.h"
 #include "theory/quantifiers/term_util.h"
 #include "options/quantifiers_options.h"
@@ -1089,7 +1090,7 @@ void TheorySep::initializeBounds() {
     for( std::map< TypeNode, TypeNode >::iterator it = d_loc_to_data_type.begin(); it != d_loc_to_data_type.end(); ++it ){
       TypeNode tn = it->first;
       Trace("sep-bound")  << "Initialize bounds for " << tn << "..." << std::endl;
-      QuantEPR * qepr = getLogicInfo().isQuantified() ? getQuantifiersEngine()->getQuantEPR() : NULL;
+      quantifiers::QuantEPR * qepr = getLogicInfo().isQuantified() ? getQuantifiersEngine()->getQuantEPR() : NULL;
       //if pto had free variable reference      
       if( d_bound_kind[tn]==bound_herbrand ){
         //include Herbrand universe of tn
